memory gauges now all 64-bit (and in bytes)

+++ trunk/src/Linux/readMemoryCounters.c
@@ -29,13 +29,12 @@ extern "C" {
       while(fgets(line, MAX_PROC_LINE_CHARS, procFile)) {
 	if(sscanf(line, "%s %"SCNu64"", var, &val64) == 2) {
 	  gotData = YES;
-	  if(strcmp(var, "MemTotal:") == 0) mem->mem_total = (uint32_t)val64;
-	  else if(strcmp(var, "MemFree:") == 0) mem->mem_free = (uint32_t)val64;
-	  // shared? $$$
-	  else if(strcmp(var, "Buffers:") == 0) mem->mem_buffers = (uint32_t)val64;
-	  else if(strcmp(var, "Cached:") == 0) mem->mem_cached = (uint32_t)val64;
-	  else if(strcmp(var, "SwapTotal:") == 0) mem->swap_total = (uint32_t)val64;
-	  else if(strcmp(var, "SwapFree:") == 0) mem->swap_free = (uint32_t)val64;
+	  if(strcmp(var, "MemTotal:") == 0) mem->mem_total = val64;
+	  else if(strcmp(var, "MemFree:") == 0) mem->mem_free = val64;
+	  else if(strcmp(var, "Buffers:") == 0) mem->mem_buffers = val64;
+	  else if(strcmp(var, "Cached:") == 0) mem->mem_cached = val64;
+	  else if(strcmp(var, "SwapTotal:") == 0) mem->swap_total = val64;
+	  else if(strcmp(var, "SwapFree:") == 0) mem->swap_free = val64;
 	}
       }
       fclose(procFile);
@@ -46,7 +45,6 @@ extern "C" {
       while(fgets(line, MAX_PROC_LINE_CHARS, procFile)) {
 	if(sscanf(line, "%s %"SCNu64"", var, &val64) == 2) {
 	  gotData = YES;
-	  // $$$ total, cached, swap, active, inactive...
 	  if(strcmp(var, "pgpgin") == 0) mem->page_in = (uint32_t)val64;
 	  else if(strcmp(var, "pgpgout") == 0) mem->page_out = (uint32_t)val64;
 	  else if(strcmp(var, "pswpin") == 0) mem->swap_in = (uint32_t)val64;
